refactor: 카드 저장의 책임 분리
